prop: letting SurfaceFlinger latch all unsignaled buffers again
Android 13 adds a new configuration called AutoSingleLayer for latching unsignaled buffers.
This configuration lets SurfaceFlinger latch an unsignaled buffer when only a single layer is updating, and not for the cases that occur across layers.
When moving to A13 the old prop was not being used in favour for `debug.sf.auto_latch_unsignaled`.

Although it has its uses we found it's best to use the old implementation for the time being regardless of the jank it adds.

Signed-off-by: itsHanibee <itshanibee@gmail.com>